Simple Example Route
Commiting this to blog about it later. Finally, after a lot of time
tinkering with WordPress and battling against trying to hijack pages
I've finally figured out a way to include a standalone php page as the
content for a page.

The best part? You can put short codes into the html and WordPress will
render them. This is because the_content hook is ... hooked, and then
the contents of the included php page are grabbed with the output
buffering functions then passed to be filtered by WordPress itself.

This little test of it works, install the plugin to your WordPress site.
Activate the plugin and direct yourself to yoursiteurl/?route=simple and
you'll see the simple example page show up. If you have the shortcodes
plugin I wrote last night installed you'll also see the wrap shortcode
in action from the included page.